NBA players of the week
Started with exploring the data and then found the average height, weight, and age of players. Then looked at who the shortest and tallest players in the league are. After that I looked at teams where players stayed on average more than 6 season. Finally I calculated the BMI for 10 random players and included the status of their weight. For a bonus I selected players who are over the age of 35 and have played more than 15 seasons in the NBA.